lots of changes to the anime service, related orms, and dtos
I was running into multiple issues with the naming conventions that made
this process more complicated than it needed to, but I learned that I
need to stick to a standard with the way I name entities to avoid this
in the future. Anyways, I worked it out and was able to populate all
tables related to the anime table, but in doing so I found an issue that
I will be fixing next.